Fix build warnings for aarch64 compatibility
- Fixed format specifier mismatches for uint64_t values in emulator.c and piscsi.c
- Fixed signed/unsigned comparison warnings by changing int to uint32_t in amiga-autoconf.c
- Fixed pointer-to-integer cast warnings in RTG code using offsetof() macro
- Added stddef.h include for offsetof() macro
- Fixed pointer casting in RTG code using uintptr_t
- Removed unused variables in AHI code
- Maintains compatibility with both 32-bit and 64-bit builds

+++ platforms/amiga/piscsi/piscsi.c
@@ -386,7 +386,7 @@ void piscsi_map_drive(char *filename, uint8_t index) {
     d->fs = file_size;
     d->fd = tmp_fd;
     lseek(tmp_fd, 0, SEEK_SET);
-    printf("[PISCSI] Map %d: [%s] - %llu bytes.\n", index, filename, file_size);
+    printf("[PISCSI] Map %d: [%s] - %lu bytes.\n", index, filename, (unsigned long)file_size);
 
     if (piscsi_parse_rdb(d) == -1) {
         DEBUG("[PISCSI] No RDB found on disk, making up some CHS values.\n");
